requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
Faker==33.3.0
iniconfig==2.0.0
kombu==5.4.2
model-bakery==1.20.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13